requirements.txt
aniso8601==6.0.0
apns2==0.5.0
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
blinker==1.4
boto3==1.9.134
botocore==1.12.134
bottlenose==1.1.8
bs4==0.0.1
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
configparser==3.7.4
cryptography==2.6.1
docutils==0.14
entrypoints==0.3
extraction==0.3
flake8==3.7.8
Flask==1.0.2
Flask-BabelEx==0.9.3
flask-compress==1.4.0
Flask-Cors==3.0.8
Flask-HTTPAuth==3.2.4
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
flask-restplus==0.12.1
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gevent==1.4.0
gfycat==0.1.4
google-api-python-client==1.7.9
Flask-Debug-API==0.1.0
Flask-DebugToolbar==0.10.1
Flask-DebugToolbar-LineProfilerPanel==1.0.2
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
greenlet==0.4.15
h2==2.6.2
hpack==3.0.0
html2text==2018.1.9
html5lib==1.0.1
httplib2==0.13.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
limits==1.3
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
oauth2client==4.1.3
oauthlib==3.0.2
passlib==1.7.1
Pillow==6.0.0
pushjack==1.6.0
pyapns-client==1.0.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.14.11
python-amazon-simple-product-api==2.2.11
python-dateutil==2.8.0
python-engineio==3.6.0
pytz==2018.9
requests==2.21.0
requests-file==1.4.3
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
six==1.12.0
soupsieve==1.9.1
speaklater==1.3
sqlakeyset==0.1.1485813522
SQLAlchemy==1.3.2
tldextract==2.2.1
uritemplate==3.0.0
urllib3==1.24.2
webencodings==0.5.1
Werkzeug==0.15.2
WTForms==2.2.1
sentry-sdk[flask]==0.13.5
mailchimp3==3.0.13
python-amazon-paapi==3.1.0